318 - Fixes dependency array warning
