+++ packages/orchestrator-ui-components/src/components/WFOWorkflowSteps/stepListUtils.ts
@@ -1,4 +1,5 @@
-import { StepState } from '../../types';
+import { Step, StepState } from '../../types';
+import { StepListItem } from './WFOStepList';
 
 export const STEP_STATE_HIDDEN_KEYS = [
     'label_',
@@ -26,4 +27,30 @@ export const getStepContent = (
             }),
             {},
         );
+};
+
+export const updateStepListItems = (
+    previousStepListItems: StepListItem[],
+    updatedSteps: Step[],
+) => {
+    const currentExpandedStateByStepId = previousStepListItems.reduce<
+        Map<string, boolean>
+    >(
+        (previousValue, currentValue) =>
+            previousValue.set(
+                currentValue.step.stepId,
+                currentValue.isExpanded,
+            ),
+        new Map(),
+    );
+
+    return updatedSteps.map((step) => {
+        const isCurrentlyExpanded = currentExpandedStateByStepId.get(
+            step.stepId,
+        );
+        return {
+            step,
+            isExpanded: isCurrentlyExpanded ?? false,
+        };
+    });
 };
